Refactor Hospitality Hub masonry
